requirements.txt
# -----------------------------
# Core App
# -----------------------------
streamlit>=1.32.0
python-dotenv>=1.0.0
PyYAML>=6.0

# -----------------------------
# OpenAI + LangChain
# -----------------------------
openai>=1.12.0
langchain>=0.1.16
langchain-openai>=0.1.7
langchain-core>=0.1.46

# -----------------------------
# Vector Database (Pinecone)
# -----------------------------
pinecone-client>=3.0.0
langchain-pinecone>=0.1.2

# -----------------------------
# Google Sheets (Supermetrics)
# -----------------------------
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# -----------------------------
# Utilities
# -----------------------------
tqdm>=4.66.0
requests>=2.31.0

# -----------------------------
# Optional (only if chat history → S3)
# -----------------------------
boto3>=1.34.0